% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emisja.R
\name{emisja}
\alias{emisja}
\title{Funckja oblicza emisję splin dla wybranej kategorii oraz standardu}
\usage{
emisja(dane, kategoria, euro, mode, substancja)
}
\arguments{
\item{dane}{dataframe}

\item{kategoria}{string}

\item{euro}{string}

\item{mode}{string}

\item{substancja}{string}
}
\value{
dataframe
}
\description{
Funckja oblicza emisję splin dla wybranej kategorii oraz standardu
}
\examples{
emisja(input, 'Passenger Cars', 'Euro 5', '', 'CO')

}
