% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\docType{data}
\name{input}
\alias{input}
\title{Dane wejściowe}
\format{
Ramka danych zawiera:
\describe{
 \item{Nat}{numeric, zawiera informacje o typie samochody np. Passenger Cars}
 \item{Segment}{character, klasa samochodu np. mini}
 \item{Paliwo}{character, rodzaj paliwa, np. petrol}
 \item{Technologia}{character, technologia użyta w silniku np. GDI}
}
}
\usage{
input
}
\description{
Dane wejściowe
}
\examples{
\dontrun{
 input
}
}
\keyword{datasets}
