% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPFuels.R
\docType{package}
\name{KPFuels}
\alias{KPFuels}
\title{KPFuels: Pakiet służący do obliczasnia emisji}
\description{
Pakiet służy do obliczania emisji zanieczyszczeń generowanych przez pojazdy
z silnikami spalinowymi i wyswietlania wyników na wykresach.
Dane są udostępniane przez Europejską Agencję Środowiska - EEA
\url{https://www.eea.europa.eu/}
}
\details{
Pakiet zawiera dwie funkcje: emisja(), emisja_plot()

\code{emisja()} Obliczanie emisji

\code{emisja_plot()} generowanie wykresu
}
