% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wskazniki}
\alias{wskazniki}
\title{Wskazniki emisji transportu drogowego EEA}
\format{

}
\source{
Europejska Agencja Środowiska \url{https://www.eea.europa.eu/themes/air/air-pollution-sources-1/emep-eea-air-pollutant-emission-inventory-guidebook}
}
\usage{
wskazniki
}
\description{
Wskazniki emisji transportu drogowego EEA
}
\examples{
\dontrun{
 wskazniki
}
}
\keyword{datasets}
