% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emisja.R
\name{emisja}
\alias{emisja}
\title{Funckja do obliczania genrowanych emisji przez pojazdy spalinowe}
\usage{
emisja(
  dane = input,
  kategoria = "Passanger Cars",
  euro = "Euro 4",
  mode = "",
  substancja = "CO"
)
}
\arguments{
\item{dane}{=    input}

\item{kategoria}{=    "Passanger Cars"}

\item{euro}{=    "Euro IV"}

\item{mode}{=    ""}

\item{substancja}{=    "CO"}
}
\value{
dataframe
}
\description{
Funckja do obliczania genrowanych emisji przez pojazdy spalinowe
}
\details{
Wzór wykorzystywany do obliczenia emsji

(Alpha x Procent^2 + Beta x Procent + Gamma + (Delta/Procent)/
(Epsilon x Procent^2 + Zita x Procent + Hta) x (1-Reduction)

Wartości domyślne funkcji
}
