% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emisja.R
\name{emisja}
\alias{emisja}
\title{Funckja do obliczania genrowanych emisji przez pojazdy spalinowe}
\usage{
emisja(
  dane = input,
  kategoria = "Passanger Cars",
  euro = "Euro 4",
  mode = "",
  substancja = "CO"
)
}
\arguments{
\item{dane}{dataframe - dane wejściowe}

\item{kategoria}{character - kategoria pojazdu np. Passenger Cars, Heavy Duty Trucks itd.}

\item{euro}{character - norma emisji spalin np. Euro II, Euro III itd.}

\item{mode}{character - tryb jazdy pojazdu np. Highway, Urban Peak itd.}

\item{substancja}{character - nazwa substacji emisyjnej np. CH4, N2O}
}
\value{
dataframe
}
\description{
Funckja do obliczania genrowanych emisji przez pojazdy spalinowe
}
\details{
Wzór wykorzystywany do obliczenia emsji

(Alpha x Procent^2 + Beta x Procent + Gamma + (Delta/Procent)/
(Epsilon x Procent^2 + Zita x Procent + Hta) x (1-Reduction)
}
