% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emisja_plot.R
\name{emisja_plot}
\alias{emisja_plot}
\title{Funkcja wykonuje wykres na podstawie danych wynikowych z funkcji "emisja"}
\usage{
emisja_plot(data)
}
\arguments{
\item{data}{- dataframe, dane wynikowe funckji "emisja"}
}
\value{
plot
}
\description{
Argumenty funkcji:
}
